+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Set logger [#80](https://github.com/rokwire/notifications-building-block/issues/80)
 
+### Changed
+- Use auth library authorization [#75](https://github.com/rokwire/notifications-building-block/issues/75)
+
 ## [1.4.0] - 2022-11-16
 ### Fixed
 - Fix docs path and the Dockerfile [#104](https://github.com/rokwire/notifications-building-block/issues/104)